requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==2.0.1
gunicorn==20.0.4
werkzeug==2.1.2
alembic==1.7.7
Faker==8.1.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.9
SQLAlchemy==1.4.22
python-decouple==3.4
python-dotenv==0.19.1